+++ APM-Final-v16/src/app/app.routes.ts
@@ -1,0 +1,8 @@
+import { Routes } from '@angular/router';
+import { WelcomeComponent } from './home/welcome.component';
+
+export const routes: Routes = [
+  { path: 'welcome', component: WelcomeComponent },
+  { path: '', redirectTo: 'welcome', pathMatch: 'full' },
+  { path: '**', redirectTo: 'welcome', pathMatch: 'full' }
+];