+++ APM-Final-v16/src/app/home/welcome.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+    templateUrl: './welcome.component.html',
+    standalone: true
+})
+export class WelcomeComponent {
+  public pageTitle = 'Welcome';
+}