+++ APM-Final-v16/src/app/products/product-detail.component.ts
@@ -1,0 +1,42 @@
+import { Component, OnInit } from '@angular/core';
+import { ActivatedRoute, Router } from '@angular/router';
+import { IProduct } from './product';
+import { ProductService } from './product.service';
+import { ConvertToSpacesPipe } from '../shared/convert-to-spaces.pipe';
+import { StarComponent } from '../shared/star.component';
+import { NgIf, LowerCasePipe, CurrencyPipe } from '@angular/common';
+
+@Component({
+    templateUrl: './product-detail.component.html',
+    styleUrls: ['./product-detail.component.css'],
+    standalone: true,
+    imports: [NgIf, StarComponent, LowerCasePipe, CurrencyPipe, ConvertToSpacesPipe]
+})
+export class ProductDetailComponent implements OnInit {
+  pageTitle = 'Product Detail';
+  errorMessage = '';
+  product: IProduct | undefined;
+
+  constructor(private route: ActivatedRoute,
+              private router: Router,
+              private productService: ProductService) {
+  }
+
+  ngOnInit(): void {
+    const id = Number(this.route.snapshot.paramMap.get('id'));
+    if (id) {
+      this.getProduct(id);
+    }
+  }
+
+  getProduct(id: number): void {
+    this.productService.getProduct(id).subscribe({
+      next: product => this.product = product,
+      error: err => this.errorMessage = err
+    });
+  }
+
+  onBack(): void {
+    this.router.navigate(['/products']);
+  }
+}