+++ APM-Final-v16/src/app/products/product-detail.guard.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { ProductDetailGuard } from './product-detail.guard';
+
+describe('ProductDetailGuard', () => {
+  let guard: ProductDetailGuard;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    guard = TestBed.inject(ProductDetailGuard);
+  });
+
+  it('should be created', () => {
+    expect(guard).toBeTruthy();
+  });
+});