+++ APM-Final-v16/src/app/shared/convert-to-spaces.pipe.ts
@@ -1,0 +1,12 @@
+import { Pipe, PipeTransform } from "@angular/core";
+
+@Pipe({
+    name: 'convertToSpaces',
+    standalone: true
+})
+export class ConvertToSpacesPipe implements PipeTransform {
+
+  transform(value: string, character: string): string {
+    return value.replace(character, ' ');
+  }
+}